Rename BulkUpdateTextInput to generic BulkUpdate
We'll collect other bulk inputs here
