+++ ui/v2.5/src/components/Studios/EditStudiosDialog.tsx
@@ -13,7 +13,7 @@ import {
   getAggregateStateObject,
 } from "src/utils/bulkUpdate";
 import { IndeterminateCheckbox } from "../Shared/IndeterminateCheckbox";
-import { BulkUpdateTextInput } from "../Shared/BulkUpdateTextInput";
+import { BulkUpdateTextInput } from "../Shared/BulkUpdate";
 import { faPencilAlt } from "@fortawesome/free-solid-svg-icons";
 import * as FormUtils from "src/utils/form";
 import { StudioSelect } from "../Shared/Select";
